+++ tt_metal/impl/debug/dprint_server.cpp
@@ -22,10 +22,12 @@
 #include "dprint_server.hpp"
 #include "debug_helpers.hpp"
 #include <rtoptions.hpp>
-#include <bfloat8.hpp>
+
+#include <tt-metalium/bfloat8.hpp>
+#include <tt-metalium/blockfloat_common.hpp>
+#include <tt-metalium/device.hpp>
 
 #include "hostdevcommon/dprint_common.h"
-#include <device.hpp>
 
 using std::cout;
 using std::endl;
